# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transcribe API",
    "version": "0.1.0",
    "description": "API contract for requesting subtitle generation and receiving subtitle-focused generated tracks."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8001",
      "description": "Local Laravel development server"
    }
  ],
  "paths": {
    "/v1/subtitle-audio/prefetch": {
      "post": {
        "operationId": "prefetchSubtitleAudio",
        "summary": "Prepare short-lived instance-local YouTube metadata without creating a generation. Disabled by default; best effort.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExtensionInstallId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "youtubeVideoId"
                ],
                "properties": {
                  "youtubeVideoId": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{11}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Best-effort prefetch queued when enabled. Does not guarantee a cache hit or start provider work.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "422": {
            "description": "Invalid YouTube video ID."
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ExtensionInstallId": {
        "name": "X-Extension-Install-Id",
        "in": "header",
        "required": true,
        "description": "Browser installation identifier for local request tracing. It is not an authentication credential or an ownership boundary.",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_-]+$"
        }
      }
    },
    "responses": {
...
      "Forbidden": {
        "description": "This request is not allowed on this private instance.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./schemas/api-error.schema.json"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./schemas/api-error.schema.json"
            }
          }
        }
      },
...
    }
  }
}